moved algorithms no longer in use to a legacy folder for reference, changed setup of main() to reduce confusion, renamed globalvars to config for clarity, moved drawObstaclesPath outside of load_files() function for modularity's sake
