+++ retired algorithms/fanpath.py
