+++ retired algorithms/rubberpath.py
